将millis()和micros()接口改为uint32_t类型

+++ ebox/core/ebox_core.h
@@ -37,8 +37,8 @@ extern "C" {
     extern void        (*interrupts)(void);
     extern int         (*no_interrupts)(void);
     extern void        (*ebox_reset)();
-    extern uint64_t    (*micros)();
-    extern uint64_t    (*millis)();
+    extern uint32_t    (*micros)();
+    extern uint32_t    (*millis)();
     extern void        (*delay_ms)(uint32_t ms);
     extern void        (*delay_us)(uint32_t us);
     extern size_t       ebox_printf(const char *fmt, ...);
